Fixed missing default theme value to default to 'light'. (#1159)
When retrieving a theme value from the component, a default value may be provided to be returned if the component does not have it's theme value set (in case of incorrectly programmatically-created components). 

In case where a default value is not provided, this leads to empty theme value being propagated to descendant components.